fill in compact element
