+++ sxt/curve21/type/compact_element.cc
@@ -1,0 +1,1 @@
+#include "sxt/curve21/type/compact_element.h"
