--- sxt/curve21/type/element_affine.cc
@@ -1,1 +1,0 @@
-#include "sxt/curve21/type/element_affine.h"
